First pass at supporting a print command.
