+++ agents_playground/terminal/terminal_action.py
@@ -1,0 +1,8 @@
+from enum import Enum, auto
+
+class TerminalAction(Enum):
+  DO_NOTHING      = auto()
+  CLOSE_TERMINAL  = auto()
+  TYPE            = auto()
+  DELETE          = auto()
+  RUN             = auto()
